Rework how resources are handled, split into many modules
