+++ nyuki/__init__.py
@@ -1,4 +1,4 @@
-from nyuki.api import Response
-from nyuki.bus import reporting
-from nyuki.capabilities import resource
-from nyuki.nyuki import Nyuki
+from .api.capabilities import resource
+from .api.webserver import Response
+from .bus import reporting
+from .nyuki import Nyuki
