+++ nyuki/api/websocket.py
@@ -1,0 +1,15 @@
+from .capabilities import resource
+from .webserver import Response
+
+
+@resource('/websocket', versions=['v1'])
+class APIWebsocketToken:
+
+    def get(self, request):
+        try:
+            self.nyuki._services.get('websocket')
+        except KeyError:
+            return Response(status=404)
+
+        token = self.nyuki.websocket.new_token()
+        return Response({'token': token})
