+++ nyuki/workflow/workflow.py
@@ -1,30 +1,29 @@
 import asyncio
-from datetime import datetime
-import json
 import logging
-from tukio import Engine, TaskRegistry, get_broker, EXEC_TOPIC
+from tukio import Engine, TaskRegistry
 from tukio.workflow import (
     TemplateGraphError, Workflow, WorkflowTemplate, WorkflowExecState
 )
-from uuid import uuid4
 
-from nyuki import Nyuki, resource, Response
+from nyuki import Nyuki
 from nyuki.bus import reporting
 
-from .storage import MongoStorage, DuplicateTemplateError
+from .api.factory import (
+    ApiFactoryRegex, ApiFactoryRegexes, ApiFactoryLookup, ApiFactoryLookups
+)
+from .api.templates import (
+    ApiTasks, ApiTemplates, ApiTemplate, ApiTemplateVersion, ApiTemplateDraft
+)
+from .api.workflows import ApiWorkflow, ApiWorkflows, ApiTest, ApiTestTopic
+from .storage import MongoStorage
 from .tasks import *
 from .tasks.utils import runtime
-from .validation import validate, TemplateError
 
 
 log = logging.getLogger(__name__)
 
 
 class BadRequestError(Exception):
-    pass
-
-
-class ConflictError(Exception):
     pass
 
 
@@ -54,6 +53,21 @@ class WorkflowNyuki(Nyuki):
             }
         }
     }
+    ENDPOINTS = Nyuki.ENDPOINTS + [
+        ApiTasks,
+        ApiTemplate,
+        ApiTemplates,
+        ApiTemplateDraft,
+        ApiTemplateVersion,
+        ApiWorkflow,
+        ApiWorkflows,
+        ApiFactoryRegex,
+        ApiFactoryRegexes,
+        ApiFactoryLookup,
+        ApiFactoryLookups,
+        ApiTest,
+        ApiTestTopic,
+    ]
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -112,612 +126,3 @@ class WorkflowNyuki(Nyuki):
             except Exception as exc:
                 # Means a bad workflow is in database, report it
                 reporting.exception(exc)
-
-    async def update_draft(self, template, from_request=None):
-        """
-        Helper to insert/update a draft
-        """
-        tmpl_dict = template.as_dict()
-
-        # Auto-increment version, draft only
-        last_version = await self.storage.templates.get_last_version(template.uid)
-        tmpl_dict['version'] = last_version + 1
-        tmpl_dict['draft'] = True
-
-        # Store task extra info (ie. title)
-        if from_request is not None:
-            rqst_tasks = from_request.get('tasks', [])
-            tmpl_tasks = tmpl_dict['tasks']
-            for src in rqst_tasks:
-                match = list(filter(lambda t: t['id'] == src['id'], tmpl_tasks))
-                if match:
-                    match[0].update({'title': src.get('title')})
-
-        try:
-            await self.storage.templates.insert_draft(tmpl_dict)
-        except DuplicateTemplateError as exc:
-            raise ConflictError('Template already exists for this version') from exc
-
-        return tmpl_dict
-
-    def errors_from_validation(self, template):
-        """
-        Validate and return the list of errors if any
-        """
-        errors = None
-        try:
-            validate(template)
-        except TemplateError as err:
-            errors = err.as_dict()
-        return errors
-
-    @resource('/workflow/templates', version='v1')
-    class Templates:
-
-        async def get(self, request):
-            """
-            Return available workflows' DAGs
-            """
-            return Response(await self.storage.templates.get_all(
-                latest=request.GET.get('latest') in ['true', 'True'],
-                draft=request.GET.get('draft') in ['true', 'True'],
-            ))
-
-        async def put(self, request):
-            """
-            Create a workflow DAG from JSON
-            """
-            request = await request.json()
-
-            if 'id' in request:
-                if await self.storage.templates.get(request['id'], draft=True):
-                    return Response(status=409, body={
-                        'error': 'draft already exists'
-                    })
-
-            try:
-                template = WorkflowTemplate.from_dict(request)
-            except TemplateGraphError as exc:
-                return Response(status=400, body={
-                    'error': str(exc)
-                })
-
-            metadata = await self.storage.templates.get_metadata(template.uid)
-            if not metadata:
-                if 'title' not in request:
-                    return Response(status=400, body={
-                        'error': "workflow 'title' key is mandatory"
-                    })
-
-                metadata = {
-                    'id': template.uid,
-                    'title': request['title'],
-                    'description': request.get('description', ''),
-                    'tags': request.get('tags', [])
-                }
-
-                await self.storage.templates.insert_metadata(metadata)
-            else:
-                metadata = metadata[0]
-
-            try:
-                tmpl_dict = await self.update_draft(template, request)
-            except ConflictError as exc:
-                return Response(status=409, body={
-                    'error': exc
-                })
-
-            return Response({
-                **tmpl_dict,
-                **metadata,
-                'errors': self.errors_from_validation(template)
-            })
-
-    @resource('/workflow/templates/{tid}', version='v1')
-    class Template:
-
-        async def get(self, request, tid):
-            """
-            Return the latest version of the template
-            """
-            tmpl = await self.storage.templates.get(tid)
-            if not tmpl:
-                return Response(status=404)
-
-            return Response(tmpl)
-
-        async def put(self, request, tid):
-            """
-            Create a new draft for this template id
-            """
-            versions = await self.storage.templates.get(tid)
-            if not versions:
-                return Response(status=404)
-
-            for v in versions:
-                if v['draft'] is True:
-                    return Response(status=409, body={
-                        'error': 'This draft already exists'
-                    })
-
-            request = await request.json()
-
-            try:
-                # Set template ID from url
-                template = WorkflowTemplate.from_dict({**request, 'id': tid})
-            except TemplateGraphError as exc:
-                return Response(status=400, body={
-                    'error': str(exc)
-                })
-
-            try:
-                tmpl_dict = await self.update_draft(template, request)
-            except ConflictError as exc:
-                return Response(status=409, body={
-                    'error': exc
-                })
-
-            metadata = await self.storage.templates.get_metadata(template.uid)
-            metadata = metadata[0]
-
-            return Response({
-                **tmpl_dict,
-                **metadata,
-                'errors': self.errors_from_validation(template)
-            })
-
-        async def patch(self, request, tid):
-            """
-            Modify the template's metadata
-            """
-            tmpl = await self.storage.templates.get(tid)
-            if not tmpl:
-                return Response(status=404)
-
-            request = await request.json()
-
-            # Add ID, request dict cleaned in storage
-            metadata = await self.storage.templates.insert_metadata({
-                **request,
-                'id': tid
-            })
-
-            return Response(metadata)
-
-        async def delete(self, request, tid):
-            """
-            Delete the template
-            """
-            tmpl = await self.storage.templates.get(tid)
-            if not tmpl:
-                return Response(status=404)
-
-            await self.storage.templates.delete(tid)
-
-            try:
-                await self.engine.unload(tid)
-            except KeyError as exc:
-                log.debug(exc)
-
-            return Response(tmpl)
-
-    @resource('/workflow/templates/{tid}/{version:\d+}', version='v1')
-    class TemplateVersion:
-
-        async def get(self, request, tid, version):
-            """
-            Return the template's given version
-            """
-            tmpl = await self.storage.templates.get(tid, version, False)
-            if not tmpl:
-                return Response(status=404)
-
-            return Response(tmpl)
-
-        async def delete(self, request, tid, version):
-            """
-            Delete a template with given version
-            """
-            tmpl = await self.storage.templates.get(tid)
-            if not tmpl:
-                return Response(status=404)
-
-            await self.storage.templates.delete(tid, version)
-            return Response(tmpl[0])
-
-    @resource('/workflow/templates/{tid}/draft', version='v1')
-    class TemplateDraft:
-
-        async def get(self, request, tid):
-            """
-            Return the template's draft, if any
-            """
-            tmpl = await self.storage.templates.get(tid, draft=True)
-            if not tmpl:
-                return Response(status=404)
-
-            return Response(tmpl[0])
-
-        async def post(self, request, tid):
-            """
-            Publish a draft into production
-            """
-            tmpl = await self.storage.templates.get(tid, draft=True)
-            if not tmpl:
-                return Response(status=404)
-
-            draft = {
-                **tmpl[0],
-                'draft': False
-            }
-
-            try:
-                # Set template ID from url
-                template = WorkflowTemplate.from_dict({**draft, 'draft': False})
-            except TemplateGraphError as exc:
-                return Response(status=400, body={
-                    'error': str(exc)
-                })
-
-            errors = self.errors_from_validation(template)
-            if errors is not None:
-                return Response(status=400, body=errors)
-
-            await self.engine.load(template)
-            # Update draft into a new template
-            await self.storage.templates.publish_draft(tid)
-            return Response(draft)
-
-        async def patch(self, request, tid):
-            """
-            Modify the template's draft
-            """
-            tmpl = await self.storage.templates.get(tid, draft=True)
-            if not tmpl:
-                return Response(status=404)
-
-            request = await request.json()
-
-            try:
-                # Set template ID from url
-                template = WorkflowTemplate.from_dict({**request, 'id': tid})
-            except TemplateGraphError as exc:
-                return Response(status=400, body={
-                    'error': str(exc)
-                })
-
-            try:
-                tmpl_dict = await self.update_draft(template, request)
-            except ConflictError as exc:
-                return Response(status=409, body={
-                    'error': str(exc)
-                })
-
-            metadata = await self.storage.templates.get_metadata(template.uid)
-            metadata = metadata[0]
-
-            return Response({
-                **tmpl_dict,
-                **metadata,
-                'errors': self.errors_from_validation(template)
-            })
-
-        async def delete(self, request, tid):
-            """
-            Delete the template's draft
-            """
-            tmpl = await self.storage.templates.get(tid, draft=True)
-            if not tmpl:
-                return Response(status=404)
-
-            await self.storage.templates.delete(tid, draft=True)
-            return Response(tmpl[0])
-
-    def serialize_wflow_exec(self, obj):
-        """
-        JSON default serializer for datetime/isoformat
-        """
-        if isinstance(obj, datetime):
-            return obj.isoformat()
-        if isinstance(obj, Workflow):
-            return obj.report()
-        raise TypeError('obj not serializable: {}'.format(obj))
-
-    def register_async_handler(self, async_channel, wflow):
-        broker = get_broker()
-        async def exec_handler(event):
-            # Pass if event does not concern this workflow execution
-            if event.source.workflow_exec_id != wflow.uid:
-                return
-            # Publish the event's data
-            # TODO: Beware of unserializable objects
-            asyncio.ensure_future(self.bus.publish(
-                event.data, topic=async_channel
-            ))
-            # If the workflow is in a final state, unregister
-            if event.data['type'] in [
-                WorkflowExecState.end.value,
-                WorkflowExecState.error.value
-            ]:
-                broker.unregister(exec_handler, topic=EXEC_TOPIC)
-        broker.register(exec_handler, topic=EXEC_TOPIC)
-
-    @resource('/workflow/instances', version='v1')
-    class Workflows:
-
-        async def get(self, request):
-            """
-            Return workflow instances
-            """
-            return Response(
-                json.dumps(
-                    [wflow for wflow in self.engine.instances.values()],
-                    default=self.serialize_wflow_exec),
-                content_type='application/json'
-            )
-
-        async def put(self, request):
-            """
-            Start a workflow from payload:
-            {
-                "id": "template_id",
-                "draft": true/false
-            }
-            """
-            async_channel = request.headers.get('X-Surycat-Async-Channel')
-            request = await request.json()
-
-            if 'id' not in request:
-                return Response(status=400, body={
-                    'error': "Template's ID key 'id' is mandatory"
-                })
-
-            draft = request.get('draft', False)
-            templates = await self.storage.templates.get(
-                request['id'],
-                draft=draft,
-                with_metadata=False
-            )
-
-            if not templates:
-                return Response(status=404, body={
-                    'error': 'Could not find a suitable template to run'
-                })
-
-            wf_tmpl = WorkflowTemplate.from_dict(templates[0])
-            data = request.get('inputs', {})
-            if draft:
-                wflow = await self.engine.run_once(wf_tmpl, data)
-            else:
-                wflow = await self.engine.trigger(wf_tmpl.uid, data)
-
-            if wflow is None:
-                return Response(status=400, body={
-                    'error': 'Could not start any workflow from this template'
-                })
-
-            # Handle async workflow exec updates
-            if async_channel is not None:
-                self.register_async_handler(async_channel, wflow)
-
-            return Response(
-                json.dumps(wflow, default=self.serialize_wflow_exec),
-                content_type='application/json'
-            )
-
-    @resource('/workflow/instances/{iid}', version='v1')
-    class Workflow:
-
-        async def get(self, request, iid):
-            """
-            Return a workflow instance
-            """
-            wflow = self.engine.instances.get(iid)
-            if not wflow:
-                return Response(status=404)
-
-            return Response(
-                json.dumps(wflow, default=self.serialize_wflow_exec),
-                content_type='application/json'
-            )
-
-        async def post(self, request, iid):
-            """
-            Operate on a workflow instance
-            """
-
-    @resource('/workflow/tasks', version='v1')
-    class Tasks:
-
-        async def get(self, request):
-            """
-            Return the available tasks
-            """
-            return Response(self.AVAILABLE_TASKS)
-
-    @resource('/workflow/topics', version='v1')
-    class Topics:
-
-        async def get(self, request):
-            """
-            Return the list of available topics
-            """
-            return Response(self.topics)
-
-    @resource('/workflow/rules', version='v1')
-    class FactoryRules:
-
-        async def get(self, request):
-            return Response(list(FACTORY_SCHEMAS.keys()))
-
-    @resource('/workflow/regexes', version='v1')
-    class Regexes:
-
-        async def get(self, request):
-            """
-            Return the list of all regexes
-            """
-            return Response(await self.storage.regexes.get_all())
-
-        async def put(self, request):
-            """
-            Insert a new regex
-            """
-            request = await request.json()
-
-            try:
-                data = {
-                    'id': str(uuid4()),
-                    'title': request['title'],
-                    'pattern': request['pattern']
-                }
-            except KeyError as exc:
-                return Response(status=400, body={
-                    'error': 'missing parameter {}'.format(exc)
-                })
-
-            await self.storage.regexes.insert(data)
-            return Response(data)
-
-        async def delete(self, request):
-            """
-            Delete all regexes and return the list
-            """
-            rules = await self.storage.regexes.get_all()
-            await self.storage.regexes.delete()
-            return Response(rules)
-
-    @resource('/workflow/regexes/{regex_id}', version='v1')
-    class Regex:
-
-        async def get(self, request, regex_id):
-            """
-            Return the regex for id `regex_id`
-            """
-            regex = await self.storage.regexes.get(regex_id)
-            if not regex:
-                return Response(status=404)
-            return Response(regex)
-
-        async def patch(self, request, regex_id):
-            """
-            Modify an existing regex
-            """
-            regex = await self.storage.regexes.get(regex_id)
-            if not regex:
-                return Response(status=404)
-
-            request = await request.json()
-
-            data = {
-                'id': regex_id,
-                'title': request.get('title', regex['title']),
-                'pattern': request.get('pattern', regex['pattern'])
-            }
-
-            await self.storage.regexes.insert(data)
-            return Response(data)
-
-        async def delete(self, request, regex_id):
-            """
-            Delete the regex with id `regex_id`
-            """
-            regex = await self.storage.regexes.get(regex_id)
-            if not regex:
-                return Response(status=404)
-
-            await self.storage.regexes.delete(regex_id)
-            return Response(regex)
-
-    @resource('/workflow/lookups', version='v1')
-    class Lookups:
-
-        async def get(self, request):
-            """
-            Return the list of all lookups
-            """
-            return Response(await self.storage.lookups.get_all())
-
-        async def put(self, request):
-            """
-            Insert a new lookup table
-            """
-            request = await request.json()
-
-            try:
-                data = {
-                    'id': str(uuid4()),
-                    'title': request['title'],
-                    'table': request['table']
-                }
-            except KeyError as exc:
-                return Response(status=400, body={
-                    'error': 'missing parameter {}'.format(exc)
-                })
-
-            await self.storage.lookups.insert(data)
-            return Response(data)
-
-        async def delete(self, request):
-            """
-            Delete all lookups and return the list
-            """
-            lookups = await self.storage.lookups.get_all()
-            await self.storage.lookups.delete()
-            return Response(lookups)
-
-    @resource('/workflow/lookups/{lookup_id}', version='v1')
-    class Lookup:
-
-        async def get(self, request, lookup_id):
-            """
-            Return the lookup table for id `lookup_id`
-            """
-            lookup = await self.storage.lookups.get(lookup_id)
-            if not lookup:
-                return Response(status=404)
-            return Response(lookup)
-
-        async def patch(self, request, lookup_id):
-            """
-            Modify an existing lookup table
-            """
-            lookup = await self.storage.lookups.get(lookup_id)
-            if not lookup:
-                return Response(status=404)
-
-            request = await request.json()
-
-            data = {
-                'id': lookup_id,
-                'title': request.get('title', lookup['title']),
-                'table': request.get('table', lookup['table'])
-            }
-
-            await self.storage.lookups.insert(data)
-            return Response(data)
-
-        async def delete(self, request, lookup_id):
-            """
-            Delete the lookup table with id `lookup_id`
-            """
-            lookup = await self.storage.lookups.get(lookup_id)
-            if not lookup:
-                return Response(status=404)
-
-            await self.storage.lookups.delete(lookup_id)
-            return Response(lookup)
-
-    @resource('/test', version='v1')
-    class Test:
-
-        async def post(self, request):
-            # Send data to all topics
-            await self.TestTopic.post(self, request, None)
-
-    @resource('/test/{topic}', version='v1')
-    class TestTopic:
-
-        async def post(self, request, topic):
-            # Send data to the given topic
-            await self.workflow_event(topic, await request.json())
