+++ tests/capabilities_test.py
@@ -3,13 +3,13 @@ from unittest import TestCase
 from unittest.mock import Mock, patch
 
 from nyuki.api import Response
-from nyuki.capabilities import resource, Capability, Exposer
+from nyuki.capabilities import resource, Capability, Api
 
 
 class TestResourceDecorator(TestCase):
 
     def test_001_call(self):
-        @resource(endpoint='/test', version='v1')
+        @resource(endpoint='/test', versions=['v1'])
         class Test:
             pass
         self.assertEqual(Test.endpoint, '/test')
@@ -50,28 +50,28 @@ class TestResponse(TestCase):
         eq_(response.content_type, 'text/plain')
 
 
-class TestExposer(TestCase):
+class TestApi(TestCase):
 
     def setUp(self):
         loop = Mock()
         self.handler = (lambda x: x)
-        self.exposer = Exposer(loop)
+        self.api = Api(loop)
         self.capability = Capability(
             name='test',
             method='GET',
             endpoint='/test',
-            version='v1',
+            versions=['v1'],
             handler=self.handler,
             wrapper=self.handler,
         )
 
     @patch('aiohttp.web_urldispatcher.UrlDispatcher.add_route')
     def test_001_register(self, add_route):
-        self.exposer.register(self.capability)
-        self.assertRaises(ValueError, self.exposer.register, self.capability)
+        self.api.register(self.capability)
+        self.assertRaises(ValueError, self.api.register, self.capability)
         add_route.assert_called_with('GET', '/v1/test', self.handler)
 
     def test_002_find(self):
-        self.assertIsNone(self.exposer._find('hello'))
-        self.exposer.register(self.capability)
-        self.assertEqual(self.exposer._find('test'), self.capability)
+        self.assertIsNone(self.api._find('hello'))
+        self.api.register(self.capability)
+        self.assertEqual(self.api._find('test'), self.capability)
